refactor(country profiles): move to server
