+++ site/server/SiteBaker.tsx
@@ -53,7 +53,7 @@ import {
 } from "explorer/covidExplorer/CovidConstants"
 import { bakeCovidChartAndVariableMeta } from "explorer/covidExplorer/bakeCovidChartAndVariableMeta"
 import { chartExplorerRedirects } from "explorer/covidExplorer/bakeCovidExplorerRedirects"
-import { countryProfileSpecs } from "site/client/CountryProfileConstants"
+import { countryProfileSpecs } from "site/server/countryProfileProjects"
 import {
     bakeAllPublishedExplorers,
     renderCovidExplorerPage
