+++ site/server/formatting.tsx
@@ -24,7 +24,7 @@ import { LoadingIndicator } from "site/client/LoadingIndicator"
 import { PROMINENT_LINK_CLASSNAME } from "site/client/blocks/ProminentLink/ProminentLink"
 import { replaceChartIframesWithExplorerIframes } from "explorer/covidExplorer/bakeCovidExplorerRedirects"
 import { SubNavId } from "./views/SiteSubnavigation"
-import { countryProfileDefaultCountryPlaceholder } from "site/client/CountryProfileConstants"
+import { countryProfileDefaultCountryPlaceholder } from "site/server/countryProfileProjects"
 
 // A modifed FontAwesome icon
 const INTERACTIVE_ICON_SVG = `<svg aria-hidden="true" focusable="false" data-prefix="fas" data-icon="hand-pointer" class="svg-inline--fa fa-hand-pointer fa-w-14" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 617">
